requirements.txt
aiohttp==3.12.15
python-telegram-bot==21.6
python-dotenv==1.1.1
uvloop==0.21.0; platform_system != 'Windows'
bs4==0.0.2
lxml==5.3.0

